fix: use more stable way to get normal
This also fixes the escaping particles in the 2d fluid example.

The issue with the previous version is that it fails if the grid node's
position is closer to the boundary than 1.0e-5.
This must be an oversight as that should be a common occurence.

The new way is more expensive, we need to check four more points against
all colliders.